Format strings for weight and height display

+++ app/src/main/java/com/marshallaf/pokemonlookup/data/PokemonData.kt
@@ -5,6 +5,6 @@ data class PokemonData(
     val name: String,
     val imageUrl: String,
     val type: String,
-    val weight: Int,
-    val height: Int
+    val weight: Double,
+    val height: Double
 )
